added realization of playback WAV file by block(chunk) of frames for the sake of testing

+++ src/block_methods.py
@@ -40,7 +40,7 @@ class RMSFromBytes(BlockAudioMethod):
 
 
 class RMSFromArray(BlockAudioMethod):
-    def __call__(self, in_data: Union[np_int16_array, np_float32_array]) -> np.float32:
+    def __call__(self, in_data: Union[np_int16_array, np_float32_array]) -> int:
         in_data = in_data.astype(np.float32)
         return round(np.sqrt((in_data * in_data).sum() / len(in_data)))
